pyproject.toml
[tool.poetry]
name = "website"
version = "0.0.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.10"
Django = "4.0.6"
wagtail = "3.0.3"
django-environ = "0.9.0"
whitenoise = {version = "6.2.0", extras = ["brotli"]}
Pygments = "2.13.0"
beautifulsoup4 = "4.9.3"
lxml = "4.9.1"
more-itertools = "8.14.0"
requests = "2.28.1"
wagtail-generic-chooser = "0.4.1"
django-rq = "2.5.1"
django-redis = "5.2.0"
wagtail-draftail-snippet = "0.4.1"
gunicorn = "20.1.0"
psycopg2 = "2.9.3"
djangorestframework = "3.13.1"
django-htmx = "1.12.2"
wagtail-metadata = "4.0.1"
humanize = "4.4.0"
django-plausible = "0.3.0"
sentry-sdk = "1.9.8"
wagtail-favicon = {git = "https://github.com/RealOrangeOne/wagtail-favicon", rev = "fd43adb98e52b6be9f5a9270391a74ec796727ef"}
django-sri = "0.5.0"
wagtail-2fa = "1.6.0"
django-health-check = "3.17.0"
wagtail-autocomplete = "0.9.0"
Wand = "^0.6.10"
django3-cache-decorator = "^0.5.2"
django-cors-headers = "^3.13.0"
uritemplate = "^4.1.1"


[tool.poetry.group.dev.dependencies]
honcho = "1.1.0"
flake8 = "4.0.1"
isort = "5.10.1"
black = "22.6.0"
django-browser-reload = "1.6.0"
django-debug-toolbar = "3.5.0"
types-requests = "2.28.5"
mypy = "0.971"
wagtail-factories = "3.1.0"
coverage = "6.5.0"
flake8-bugbear = "22.8.23"
flake8-builtins = "1.5.3"
flake8-comprehensions = "3.10.0"
flake8-mutable = "1.2.0"
flake8-print = "5.0.0"
flake8-tuple = "0.4.1"
djlint = "1.18.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.djlint]
ignore = "T003,H006,H023,H030,H031"
indent = 2
preserve_blank_lines = true
profile = "django"
max_line_length = 12000
max_attribute_length = 1000
extension = "html"
format_attribute_template_tags=true
custom_blocks="cache"

[tool.djlint.per-file-ignores]
"_snippet_link.html" = "H025"
